+++ routingpy/routers/google.py
@@ -354,12 +354,8 @@ class Google:
                     duration += leg["duration"]["value"]
                     distance += leg["distance"]["value"]
                     for step in leg["steps"]:
-                        geometry.extend(
-                            [
-                                list(reversed(coords))
-                                for coords in utils.decode_polyline5(step["polyline"]["points"])
-                            ]
-                        )
+                        geometry.extend(utils.decode_polyline5(step["polyline"]["points"]))
+
                 routes.append(
                     Direction(
                         geometry=geometry, duration=int(duration), distance=int(distance), raw=route