+++ tests/test_utils.py
@@ -22,10 +22,10 @@ import tests as _test
 
 class UtilsTest(_test.TestCase):
     def setUp(self):
-        self.coords2d_5prec = "__`t@smslHfo@~\\"
-        self.coords3d_5prec = "__`t@smslH_sVfo@~\\etjG"
-        self.coords2d_6prec = "a_iqOaqkg}AxaN`kH"
-        self.coords3d_6prec = "a_iqOaqkg}A_sVxaN`kHetjG"
+        self.coords2d_5prec = r"smslH__`t@~\fo@"
+        self.coords3d_5prec = r"smslH__`t@_sV~\fo@etjG"
+        self.coords2d_6prec = r"aqkg}Aa_iqO`kHxaN"
+        self.coords3d_6prec = r"aqkg}Aa_iqO_sV`kHxaNetjG"
 
     def test_polyline5_2d_decoding(self):
 
@@ -46,6 +46,11 @@ class UtilsTest(_test.TestCase):
 
         decoded = [(8.688641, 49.420577, 120.96), (8.680916, 49.415776, 1491.39)]
         self.assertEqual(decoded, utils.decode_polyline6(self.coords3d_6prec, True))
+
+    def test_polyline6_3d_decoding_latlng(self):
+
+        decoded = [(49.420577, 8.688641, 120.96), (49.415776, 8.680916, 1491.39)]
+        self.assertEqual(decoded, utils.decode_polyline6(self.coords3d_6prec, True, order="latlng"))
 
     def test_get_ordinal(self):